reset main and app
