+++ client/src/main.jsx
@@ -13,13 +13,6 @@ import OneGoal from './pages/Goals/One-Goal.jsx'
 import Home from './pages/Home/Home.jsx'
 import Report from './pages/Report/Report.jsx'
 import Profile from './pages/Profile/Profile.jsx'
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from '@apollo/client'
-import { setContext } from '@apollo/client/link/context'
 
 const domain = import.meta.env.VITE_REACT_APP_AUTH0_DOMAIN
 const clientId = import.meta.env.VITE_REACT_APP_AUTH0_CLIENT_ID
@@ -76,32 +69,7 @@ const router = createBrowserRouter([
     ],
   },
 ])
-// Construct our main GraphQL API endpoint
-const httpLink = createHttpLink({
-  uri: '/graphql',
-})
-
-// Construct request middleware that will attach the JWT token to every request as an `authorization` header
-const authLink = setContext((_, { headers }) => {
-  // get the authentication token from local storage if it exists
-  const token = localStorage.getItem('id_token')
-  // return the headers to the context so httpLink can read them
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : '',
-    },
-  }
-})
-
-const client = new ApolloClient({
-  // Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-})
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <ApolloProvider client={client}>
-    <RouterProvider router={router} />
-  </ApolloProvider>
+  <RouterProvider router={router} />
 )
